requirements.txt
alabaster==0.7.12
anaconda-client==1.9.0
anaconda-navigator==2.1.1
anaconda-project==0.10.2
anyio==3.5.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==0.13.1
asn1crypto==1.4.0
astroid==2.6.6
astropy==5.0.2
asttokens==2.0.5
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
bitarray==2.3.5
bkcharts==0.2
black==19.10b0
bleach==4.1.0
bokeh==2.4.2
boto==2.49.0
Bottleneck==1.3.4
brotlipy==0.7.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.0.0
clyent==1.2.2
colorama==0.4.4
comtypes==1.1.10
conda==4.12.0
conda-build==3.21.6
conda-content-trust==0+unknown
conda-pack==0.6.0
conda-package-handling==1.8.0
conda-repo-cli==1.0.4
conda-token==0.3.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
cookiecutter==1.7.2
cryptography==3.4.8
cycler==0.11.0
Cython==0.29.28
cytoolz==0.11.0
daal4py==2021.5.0
dask==2022.2.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distributed==2022.2.1
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.3
fastcache==1.1.0
filelock==3.6.0
flake8==3.9.2
Flask==1.1.2
fonttools==4.25.0
fsspec==2022.2.0
future==0.18.2
gevent==21.8.0
glob2==0.7
greenlet==1.1.1
h5py==3.6.0
HeapDict==1.0.1
html5lib==1.1
idna==3.3
imagecodecs==2021.8.26
imageio==2.9.0
imagesize==1.3.0
importlib-metadata==4.8.2
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.1
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.0
keyring==23.4.0
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
libarchive-c==2.9
llvmlite==0.37.0
locket==0.2.1
lxml==4.8.0
MarkupSafe==1.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.2
mccabe==0.6.1
menuinst==1.4.18
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mock==4.0.3
more-itertools==8.12.0
mpmath==1.2.1
msgpack==1.0.2
multipledispatch==0.6.0
munkres==1.1.4
mypy-extensions==0.4.3
navigator-updater==0.2.1
nbclassic==0.3.5
nbclient==0.5.11
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.7.1
nltk==3.7
nose==1.3.7
notebook==6.4.8
numba==0.54.1
numexpr==2.8.1
numpy==1.20.3
numpydoc==1.2
olefile==0.46
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
paramiko==2.8.1
parso==0.8.3
partd==1.2.0
path==16.2.0
pathlib2==2.3.6
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
pkginfo==1.8.2
pluggy==1.0.0
ply==3.11
poyo==0.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.20
psutil==5.8.0
psycopg2==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.7.0
pycosat==0.6.3
pycparser==2.21
pycurl==7.44.1
pydocstyle==6.1.1
pyerfa==2.0.0
pyflakes==2.3.1
Pygments==2.11.2
PyJWT==2.1.0
pylint==2.9.6
pyls-spyder==0.4.0
PyNaCl==1.4.0
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==3.0.4
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.20.0
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-slugify==5.0.2
pytz==2021.3
PyWavelets==1.1.1
pywin32==302
pywin32-ctypes==0.2.0
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.3
qtconsole==5.2.2
QtPy==1.11.2
regex==2021.8.3
requests==2.27.1
rope==0.22.0
Rtree==0.9.7
ruamel-yaml-conda==0.15.100
scikit-image==0.18.3
scikit-learn==1.0.2
scikit-learn-intelex==2021.20220215.102710
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
simplegeneric==0.8.1
singledispatch==3.7.0
sip==4.19.13
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
spyder==5.1.5
spyder-kernels==2.1.3
SQLAlchemy==1.4.32
stack-data==0.2.0
statsmodels==0.12.2
sympy==1.10.1
tables==3.6.1
TBB==0.2
tblib==1.7.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss==0.4
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typed-ast==1.4.3
typing_extensions==4.1.1
ujson==5.1.0
unicodecsv==0.14.1
Unidecode==1.2.0
urllib3==1.26.8
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.0.3
wheel==0.37.1
whichcraft==0.6.1
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==3.0.2
xlwings==0.24.9
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.31.0
zict==2.0.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
